## ChangeLog
- Implemented an Error Handler to unify repeatingly used error messages
- Added new Localization Strings
- Implemented new String-Helper function :Contains
- set error level for error message in :CleanUp Table-Helper function

## NEW Feature

FrameSelector implemented. This is a Function which can be called by a button click to get the name of a frame currently highlighted with the mouse cursor. The frame is shown embordered and annotated with a tooltip containing the frame's name. Clicking the frame updates a given option field and fires custom OnChange event for this field.